feat: first endpoint (#58)
* feat: first endpoint

* refactor: solve issues but expand

* feat: get endpoint finished

* feature: list endpoint done

* fix: residuosArray fix

* feat: add post endpoint

* feat: put residuo endpoint

* feat: TransaccionPutFecha endpoint

* feat: add four endpoints

* all endpoints but solicitud set

* fix test 1

* fix: test ok

---------

+++ src/main/java/org/circle8/controller/response/TransporteResponse.java
@@ -11,16 +11,16 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder(toBuilder = true)
 public class TransporteResponse implements ApiResponse {
-	public int id;
+	public Long id;
 	public ZonedDateTime fechaAcordada;
 	public ZonedDateTime fechaInicio;
 	public ZonedDateTime fechaFin;
 	public BigDecimal precioAcordado;
 	public String transportistaUri;
-	public Integer transportistaId;
+	public Long transportistaId;
 	public TransportistaResponse transportista;
 	public String transaccionUri;
-	public Integer transaccionId;
+	public Long transaccionId;
 	public boolean pagoConfirmado;
 	public boolean entregaConfirmada;
